+++ packages/myst-cli/src/cli/options.ts
@@ -1,4 +1,5 @@
 import { InvalidArgumentError, Option } from 'commander';
+import { cpus } from 'node:os';
 
 export const MYST_DOI_BIB_FILE = 'myst.doi.bib';
 
@@ -60,6 +61,19 @@ export function makeExecuteOption(description: string) {
   return new Option('--execute', description).default(false);
 }
 
+export function makeExecuteParallelOption() {
+  const defaultParallelism = Math.max(1, cpus().length - 1);
+  return new Option('--execute-parallel <n>', `Maximum number of notebooks to execute in parallel`)
+    .argParser((value) => {
+      const parsedValue = parseInt(value);
+      if (parsedValue < 1) {
+        throw new InvalidArgumentError('Must be an integer greater than or equal to 1.');
+      }
+      return parsedValue;
+    })
+    .default(defaultParallelism);
+}
+
 export function makeAllOption(description: string) {
   return new Option('-a, --all', description).default(false);
 }
@@ -82,7 +96,10 @@ export function makeNamedExportOption(description: string) {
 }
 
 export function makeStrictOption() {
-  return new Option('--strict', 'Summarize build warnings and stop on any errors.').default(false);
+  return new Option(
+    '--strict',
+    'Summarize build warnings and exit non-zero on any errors.',
+  ).default(false);
 }
 
 export function makeForceOption(description: string) {
@@ -90,7 +107,10 @@ export function makeForceOption(description: string) {
 }
 
 export function makeCheckLinksOption() {
-  return new Option('--check-links', 'Check all links to websites resolve.').default(false);
+  return new Option(
+    '--check-links',
+    'Check all external links resolve and report broken ones in the build log.',
+  ).default(false);
 }
 
 export function makeKeepHostOption() {