+++ packages/myst-ext-grid/src/index.ts
@@ -50,6 +50,15 @@ export const gridDirective: DirectiveSpec = {
 
 export const gridItemDirective: DirectiveSpec = {
   name: 'grid-item',
+  // Other options from sphinx-design we leave below for reference in case we want to support in the future
+  // options:
+  // // https://sphinx-design.readthedocs.io/en/furo-theme/grids.html#grid-item-options
+  // margin
+  // padding
+  // child-direction
+  // child-align
+  // outline
+  // class
   options: {
     label: {
       type: String,
@@ -57,6 +66,9 @@ export const gridItemDirective: DirectiveSpec = {
     },
     class: {
       type: String,
+    },
+    columns: {
+      type: Number,
     },
   },
   body: {
@@ -66,7 +78,8 @@ export const gridItemDirective: DirectiveSpec = {
   run(data: DirectiveData): GenericNode[] {
     const { label, identifier } = normalizeLabel(data.options?.label as string | undefined) || {};
     const div: GenericNode = {
-      type: 'div',
+      type: 'grid-item',
+      columns: data.options?.columns as number | undefined,
       children: data.body as unknown as GenericNode[],
       class: data.options?.class as string | undefined,
       label,