+++ packages/tex-to-myst/src/figures.ts
@@ -43,7 +43,7 @@ const FIGURE_HANDLERS: Record<string, Handler> = {
   macro_centering: centering,
   macro_includegraphics(node, state) {
     state.closeParagraph();
-    const url = texToText(getArguments(node, 'group'));
+    const url = texToText(getArguments(node, 'group')).replace(/"/g, '');
     const args = getArguments(node, 'argument')?.[0]?.content ?? [];
     // TODO: better width, placement, etc.
     if (