+++ packages/markdown-it-myst/CHANGELOG.md
@@ -1,5 +1,17 @@
 # markdown-it-myst
 
+## 1.0.16
+
+### Patch Changes
+
+- 9355519: Add multi-line to role parsing
+
+## 1.0.15
+
+### Patch Changes
+
+- bc240a8: Make citation parsing less aggressive
+
 ## 1.0.14
 
 ### Patch Changes