+++ packages/myst-common/CHANGELOG.md
@@ -1,5 +1,20 @@
 # myst-common
 
+## 1.9.5
+
+### Patch Changes
+
+- 7b989d4: Cheanges to suport first version of anywidget
+- Updated dependencies [7b989d4]
+  - myst-frontmatter@1.9.5
+
+## 1.9.4
+
+### Patch Changes
+
+- Updated dependencies [045953f]
+  - myst-frontmatter@1.9.4
+
 ## 1.9.3
 
 ### Patch Changes