+++ packages/myst-execute/CHANGELOG.md
@@ -1,5 +1,17 @@
 # myst-execute
 
+## 0.3.5
+
+### Patch Changes
+
+- 72172ff: Await connection end for jupyter kernel
+
+## 0.3.4
+
+### Patch Changes
+
+- 0ddb337: Fix handling of update_display_data
+
 ## 0.3.3
 
 ### Patch Changes