+++ packages/myst-roles/CHANGELOG.md
@@ -1,5 +1,27 @@
 # myst-roles
 
+## 1.7.2
+
+## 1.7.1
+
+## 1.7.0
+
+### Patch Changes
+
+- Updated dependencies [7b989d4]
+  - myst-spec-ext@1.9.5
+  - myst-common@1.9.5
+
+## 1.6.5
+
+## 1.6.4
+
+### Patch Changes
+
+- 045953f: Add support for parsing decimal numbers in si role
+  - myst-common@1.9.4
+  - myst-spec-ext@1.9.4
+
 ## 1.6.3
 
 ### Patch Changes