+++ packages/myst-to-html/CHANGELOG.md
@@ -1,5 +1,24 @@
 # myst-to-html
 
+## 1.7.2
+
+## 1.7.1
+
+## 1.7.0
+
+### Patch Changes
+
+- Updated dependencies [7b989d4]
+  - myst-common@1.9.5
+
+## 1.6.5
+
+## 1.6.4
+
+### Patch Changes
+
+- myst-common@1.9.4
+
 ## 1.6.3
 
 ### Patch Changes